스터디 상세 페이지 API 연동 (참여, 상세조회, 기록등록, 기록조회)

+++ src/main/java/com/example/studymate/Study/dto/AddStudyRequest.java
@@ -24,7 +24,7 @@ public class AddStudyRequest {
     @JsonFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss")
     private LocalDateTime endDate;
 
-    private Integer participantsMax;
+    private Long participantsMax;
     private LocalDateTime recruitDeadline;
